+++ Plugins/Events/Events/DMActionEvents.cpp
@@ -1,6 +1,9 @@
 #include "Events.hpp"
 #include "API/CAppManager.hpp"
 #include "API/CServerExoApp.hpp"
+#include "API/CNWSFaction.hpp"
+#include "API/CServerExoAppInternal.hpp"
+#include "API/CFactionManager.hpp"
 #include "API/CGameObjectArray.hpp"
 #include "API/CNetLayer.hpp"
 #include "API/CNetLayerPlayerInfo.hpp"
@@ -477,7 +480,49 @@ int32_t HandleDMMessageHook(CNWSMessage *thisPtr, CNWSPlayer *pPlayer, uint8_t n
         case MessageDungeonMasterMinor::SetFactionByName:
         {
             event += "SET_FACTION";
-            DefaultSignalEvent();
+
+            int32_t offset = 0;
+
+            int32_t factionid;
+            std::string factionName;
+            std::string target;
+
+            if (nMinor == MessageDungeonMasterMinor::SetFaction)
+            {
+                factionid = Utils::PeekMessage<int32_t>(thisPtr, offset); offset += sizeof(int32_t);
+                target = Utils::ObjectIDToString(Utils::PeekMessage<ObjectID>(thisPtr, offset) & 0x7FFFFFFF);
+
+                auto* pFaction = Globals::AppManager()->m_pServerExoApp->m_pcExoAppInternal->m_pFactionManager->GetFaction(factionid);
+                if (pFaction)
+                {
+                    factionName = pFaction->m_sFactionName;
+                }
+            }
+            else
+            {
+                target = Utils::ObjectIDToString(Utils::PeekMessage<ObjectID>(thisPtr, offset) & 0x7FFFFFFF); offset += sizeof(ObjectID);
+                factionName = Utils::PeekMessage<std::string>(thisPtr, offset);
+
+                factionid = Globals::AppManager()->m_pServerExoApp->m_pcExoAppInternal->m_pFactionManager->GetFactionIdByName(factionName);
+            }
+
+            auto PushAndSignal = [&](const std::string& ev) -> bool {
+                PushEventData("TARGET", target);
+                PushEventData("FACTION_ID", std::to_string(factionid));
+                PushEventData("FACTION_NAME", factionName);
+                return SignalEvent(ev, oidDM);
+            };
+
+            if (PushAndSignal(event + "_BEFORE"))
+            {
+                retVal = s_HandlePlayerToServerDungeonMasterMessageHook->CallOriginal<int32_t>(thisPtr, pPlayer, nMinor, bGroup);
+            }
+            else
+            {
+                retVal = false;
+            }
+
+            PushAndSignal(event +"_AFTER");
             break;
         }
         case MessageDungeonMasterMinor::GiveItem:
@@ -526,19 +571,108 @@ int32_t HandleDMMessageHook(CNWSMessage *thisPtr, CNWSPlayer *pPlayer, uint8_t n
         case MessageDungeonMasterMinor::SetStat:
         {
             event += "SET_STAT";
-            DefaultSignalEvent();
+
+            int32_t offset = 0;
+
+            int32_t stat = Utils::PeekMessage<int32_t>(thisPtr, offset); offset += sizeof(int32_t);
+            std::string value = std::to_string(Utils::PeekMessage<float>(thisPtr, offset)); offset += sizeof(float);
+            std::string target = Utils::ObjectIDToString(Utils::PeekMessage<ObjectID>(thisPtr, offset) & 0x7FFFFFFF); offset += sizeof(ObjectID);
+            std::string set = std::to_string((bool)(Utils::PeekMessage<int32_t>(thisPtr, offset) & 0x10));
+
+            auto PushAndSignal = [&](const std::string& ev) -> bool {
+                PushEventData("STAT", std::to_string(stat - 5));
+                PushEventData("VALUE", value);
+                PushEventData("TARGET", target);
+                PushEventData("SET", set);
+                return SignalEvent(ev, oidDM);
+            };
+
+            if (PushAndSignal(event + "_BEFORE"))
+            {
+                retVal = s_HandlePlayerToServerDungeonMasterMessageHook->CallOriginal<int32_t>(thisPtr, pPlayer, nMinor, bGroup);
+            }
+            else
+            {
+                retVal = false;
+            }
+
+            PushAndSignal(event +"_AFTER");
             break;
         }
         case MessageDungeonMasterMinor::GetVar:
         {
             event += "GET_VARIABLE";
-            DefaultSignalEvent();
+            
+            int32_t offset = 0;
+
+            uint8_t varType = Utils::PeekMessage<uint8_t>(thisPtr, offset); offset += sizeof(uint8_t);
+            std::string target = Utils::ObjectIDToString(Utils::PeekMessage<ObjectID>(thisPtr, offset) & 0x7FFFFFFF); offset += sizeof(ObjectID);
+            std::string key = Utils::PeekMessage<std::string>(thisPtr, offset);
+
+            auto PushAndSignal = [&](const std::string& ev) -> bool {
+                PushEventData("TYPE", std::to_string(varType));
+                PushEventData("TARGET", target);
+                PushEventData("KEY", key);
+                return SignalEvent(ev, oidDM);
+            };
+
+            if (PushAndSignal(event + "_BEFORE"))
+            {
+                retVal = s_HandlePlayerToServerDungeonMasterMessageHook->CallOriginal<int32_t>(thisPtr, pPlayer, nMinor, bGroup);
+            }
+            else
+            {
+                retVal = false;
+            }
+
+            PushAndSignal(event +"_AFTER");
             break;
         }
         case MessageDungeonMasterMinor::SetVar:
         {
             event += "SET_VARIABLE";
-            DefaultSignalEvent();
+
+            int32_t offset = 0;
+
+            uint8_t varType = Utils::PeekMessage<uint8_t>(thisPtr, offset); offset += sizeof(uint8_t);
+            std::string target = Utils::ObjectIDToString(Utils::PeekMessage<ObjectID>(thisPtr, offset) & 0x7FFFFFFF); offset += sizeof(ObjectID);
+            std::string key = Utils::PeekMessage<std::string>(thisPtr, offset); offset += key.length() + 4;
+            std::string value;
+
+            switch (varType)
+            {
+                case 0:
+                    value = std::to_string(Utils::PeekMessage<uint32_t>(thisPtr, offset));
+                    break;
+                case 1:
+                    value = std::to_string(Utils::PeekMessage<float>(thisPtr, offset));
+                    break;
+                case 2:
+                    value = Utils::PeekMessage<std::string>(thisPtr, offset);
+                    break;
+                case 3:
+                    value = Utils::ObjectIDToString(Utils::PeekMessage<ObjectID>(thisPtr, offset));
+                    break;
+            }
+
+            auto PushAndSignal = [&](const std::string& ev) -> bool {
+                PushEventData("TYPE", std::to_string(varType));
+                PushEventData("TARGET", target);
+                PushEventData("KEY", key);
+                PushEventData("VALUE", value);
+                return SignalEvent(ev, oidDM);
+            };
+
+            if (PushAndSignal(event + "_BEFORE"))
+            {
+                retVal = s_HandlePlayerToServerDungeonMasterMessageHook->CallOriginal<int32_t>(thisPtr, pPlayer, nMinor, bGroup);
+            }
+            else
+            {
+                retVal = false;
+            }
+
+            PushAndSignal(event +"_AFTER");
             break;
         }
         case MessageDungeonMasterMinor::SetTime: